Added SebDaBaBo's Hackpad
(I wrote @xX_ALEXREN_Xx today: Due to a not functioning PC I can't finish this project by the end of today and I would still greatly appreciate it, if you amazing people could send me the PCB and the components so I can finish this awesome project at home, cause I have a 3d printer. After I've put everything together I will post a photo in the Hack Club Slack and thank you all again!)

A few weeks ago I bought my first custom Keyboard which was a huge achievement after typing for a long time on the keyboard that came with my PC. I loved the typing experience: The feeling of pressing the Gateron Switches and the sound! Since then I always wanted to build a macropad to make my life easier when using the PC with shortcuts. I never knew where to start and I watched dozens of videos. With an opportunity like this I put in all my energy and effort to get this done on the last weekend of the event, cause my PC was malfunctioning! The Hardest challenge I encountered, was trying to learn and create everything in the 6 hours I had on the weekend. There was a lot of trouble-shooting and experimenting involved. I would like to have another Macropad/keyboard event or maybe general HID devices...