Added Production Folder and Updated Case
I realized that I do not have a production folder, so I updated the structure of the folders and location of the proper files.
Also I realized that the heat insert holes were too close to the edge of the bottom part, so fixed that as well. And my text which was originally on the bottom was flipped and upside down, so I went to fix it but eventually decided that it would be better to add some text on the top of the case.